+++ test/unit/query_spec.js
@@ -2,6 +2,7 @@ require("../test_helper")
 
 import Strategy from "../../src/strategy"
 import SearchResult from "../../src/search_result"
+import Query from "../../src/query"
 import { createQuery } from "../test_utils"
 
 const assert = require("power-assert")
@@ -13,6 +14,96 @@ describe("Query", function() {
     term = "he"
     match = [term, "", term]
     query = createQuery(undefined, term, match)
+  })
+
+  describe(".build()", function() {
+    var strategy
+
+    function sharedExamples() {
+      context("and given text matches to the function's result", function() {
+        it("should return a Query object", function() {
+          var result = Query.build(strategy, "@hello")
+          assert.ok(result instanceof Query)
+          assert.strictEqual(result.strategy, strategy)
+          assert.strictEqual(result.term, "hello")
+          assert.ok(Array.isArray(result.match))
+        })
+      })
+
+      context(
+        "and given text does not match to the function's result",
+        function() {
+          it("should return null", function() {
+            assert.equal(Query.build(strategy, "hello"), null)
+          })
+        },
+      )
+    }
+
+    context("when match param is a function", function() {
+      beforeEach(function() {
+        strategy = new Strategy({
+          match: function(text) {
+            return text.match(/(^|\s)@(\w+)$/)
+          },
+        })
+      })
+
+      sharedExamples()
+    })
+
+    context("when match param is a regexp", function() {
+      beforeEach(function() {
+        strategy = new Strategy({ match: /(^|\s)@(\w+)$/ })
+      })
+
+      sharedExamples()
+    })
+
+    context("when context prop is given", function() {
+      var props
+
+      beforeEach(function() {
+        props = { match: /(^|\s)@(\w+)$/ }
+        strategy = new Strategy(props)
+      })
+
+      it("should call the function", function() {
+        var spy = (props.context = this.sinon.spy())
+        var text = "@hello"
+        Query.build(strategy, text)
+        assert(spy.calledOnce)
+        assert(spy.calledWith(text))
+      })
+
+      context("and it returns false", function() {
+        beforeEach(function() {
+          props.context = this.sinon.spy(function() {
+            return false
+          })
+        })
+
+        it("should return null even if the text matches", function() {
+          assert.equal(Query.build(strategy, "@hello"), null)
+        })
+      })
+
+      context("and it returns a string", function() {
+        var returnedString
+        beforeEach(function() {
+          returnedString = "@world"
+          props.context = this.sinon.spy(function() {
+            return returnedString
+          })
+        })
+
+        it("should test match with the returned string", function() {
+          var text = "@hello"
+          var result = Query.build(strategy, text)
+          assert.strictEqual(result.term, "world")
+        })
+      })
+    })
   })
 
   describe("#strategy", function() {
